display final result in progress
Now in progress is to display the final result of each player and the dealer. The box with the result should be hidden until it is necessary to use it (maybe create it in JS or hide it first with JS).  And also the text in the box should be centered.
Next thing is to place the boxes in the right x an y coordinates (with cartesian coordinate system like the chair cards place!)
--> It would be cool, to create transparent boxes (not 100%)!